feat(image) : Image 엔티티 abstract 로 변경, 상속받아서 구체 Image 클래스 구현하도록 변경
